[validation] validation strategy
