--- src/strategies/ticket-multiple-validity/README.md
@@ -1,11 +1,0 @@
-# Ticket multiple validity strategy
-
-This strategy is improved from `ticket-validity` strategy, It checks if the user has minimum voting power from at least one strategy in the strategies list
-
-Parameters:
-
-| Parameter     | Description                                  | Default value |
-| ------------- | ------------------------------------------   | ------------- |
-| `strategies`  | List of strategies (Max 3)                   |               |
-| `min`         | Minimum voting power in a strategy           | 0             |
-| `symbol`      | Token symbol                                 | optional      |
